chore: move headers to model folder
