+++ src/plugins/comments/model/localizationresourcereference.h
